wip: repair export, change batch mode to account for WebView2, repair airbase radio checkbox edition, start work on map drawings fine tuning

+++ Tools/ToolsMap.cs
@@ -104,7 +104,7 @@ namespace DcsBriefop.Tools
 			float fFontSize = 11;
 			if (drawingObject.FontSize is not null)
 			{
-				fFontSize = (float)(drawingObject.FontSize.Value - 3);
+				fFontSize = drawingObject.FontSize.Value - 3;
 				if (fFontSize < 1)
 					fFontSize = 1;
 			}
